cxl: qemu: Test volatile memory with LSA
Link: https://www.qemu.org/docs/master/system/devices/cxl.html

+++ cxl/vm.sh
@@ -3,11 +3,13 @@ set -e
 
 sudo dnf install -y cxl-cli
 
-sudo dracut --kver $(uname -r) --no-hostonly --verbose --force \
-	--install 'insmod rmmod modprobe lspci ndctl cxl' \
-	--add-drivers 'cxl_acpi cxl_core cxl_mem cxl_pci cxl_pmem cxl_pmu cxl_port' \
-	initramfs.img
+if ! [[ -e initramfs.img ]]; then
+	sudo dracut --kver $(uname -r) --no-hostonly --verbose --force \
+		--install 'insmod rmmod modprobe lspci ndctl cxl' \
+		--add-drivers 'cxl_acpi cxl_core cxl_mem cxl_pci cxl_pmem cxl_pmu cxl_port' \
+		initramfs.img
+fi
 
-sudo cp /boot/vmlinuz-$(uname -r) vmlinuz
+[[ ! -e vmlinuz ]] && sudo cp /boot/vmlinuz-$(uname -r) vmlinuz
 
-sudo ../init/rootfs/qemu.sh -k vmlinuz -r initramfs.img --initrd --cxl cxl-pmem --stdio
+sudo ../init/rootfs/qemu.sh -k vmlinuz -r initramfs.img --initrd --cxl cxl-vmem-lsa --stdio
